Move card components to one directory
