+++ lib/KThemePlugin.js
@@ -5,8 +5,8 @@ import { darken1, darken2, darken3 } from './styles/darkenColors.js';
 import KBreadcrumbs from './KBreadcrumbs';
 import KButton from './buttons-and-links/KButton';
 import KButtonGroup from './buttons-and-links/KButtonGroup';
-import KCard from './KCard';
-import KCardGrid from './KCardGrid';
+import KCard from './cards/KCard';
+import KCardGrid from './cards/KCardGrid';
 import KCheckbox from './KCheckbox';
 import KCircularLoader from './loaders/KCircularLoader';
 import KDateRange from './KDateRange';
