+++ lib/cards/__tests__/KCard.spec.js
@@ -1,6 +1,6 @@
 import { mount, createLocalVue } from '@vue/test-utils';
 import VueRouter from 'vue-router';
-import KCard from './../index.vue';
+import KCard from '../KCard';
 
 const localVue = createLocalVue();
 localVue.use(VueRouter);
